SuperDesktopTextField scroll tests moved to separate file, refactor
