+++ super_editor/example/lib/demos/supertextfield/textfield_inside_slivers_demo.dart
@@ -29,9 +29,8 @@ class _TextFieldInsideSliversDemoState extends State<TextFieldInsideSliversDemo>
           child: Padding(
             padding: const EdgeInsets.all(8.0),
             child: SuperDesktopTextField(
-              // Provides sufficient space to test out textfields scrolling behaviour
+              minLines: 5, // Provides sufficient space to test out textfields scrolling behaviour
               // properly.
-              minLines: 5,
               maxLines: 5,
               textStyleBuilder: demoTextStyleBuilder,
               decorationBuilder: (context, child) {
